refactor color render code
